refactor(web): move nextAvailableSlug into util/slug helper
Share one slug-generation implementation between the create form and the
reuse flow; copyReuse re-exports it so existing call sites are unchanged.
